fixed PR comments.
